feat: domain exception 관련 코드 domain 패키지 하위로 이동
